feat: Remove `track_count` column
Change to: get media count by DB query

+++ shared/compose/usecase/src/commonMain/kotlin/com/andannn/melodify/shared/compose/usecase/LibraryDataSourceContentUseCase.kt
@@ -90,13 +90,13 @@ fun LibraryDataSource.item(): Flow<MediaItemModel?> =
     when (this) {
         is LibraryDataSource.AlbumDetail -> {
             repository.getAlbumByAlbumIdFlow(
-                id,
+                id.toLong(),
             )
         }
 
         is LibraryDataSource.ArtistDetail -> {
             repository.getArtistByArtistIdFlow(
-                id,
+                id.toLong(),
             )
         }
